+++ tests/test_utils.py
@@ -14,7 +14,12 @@ from jax_unirep.utils import (
     load_random_evotuning_params,
     right_pad,
     validate_mLSTM1900_params,
+    evotuning_pairs,
+    input_output_pairs,
+    length_batch_input_outputs,
 )
+
+from contextlib import suppress as does_not_raise
 
 
 def test_l2_normalize():
@@ -122,3 +127,38 @@ def test_right_pad(seqs, max_len, expected):
 def test_load_random_evotuning_params():
     params = load_random_evotuning_params()
     validate_params(params)
+
+
+@pytest.mark.parametrize(
+    "seqs, expected",
+    [
+        ([], pytest.raises(ValueError)),
+        (["MT", "MTN"], pytest.raises(ValueError)),
+        (["MT", "MB", "MD"], does_not_raise()),
+    ],
+)
+def test_input_output_pairs(seqs, expected):
+
+    with expected:
+        assert input_output_pairs(seqs) is not None
+
+    if expected == does_not_raise():
+        xs, ys = input_output_pairs(seqs)
+        assert xs.shape == (len(seqs), len(seqs[0]) + 1, 10)
+        assert ys.shape == (len(seqs), len(seqs[0]) + 1, 25)
+
+
+def test_length_batch_input_outputs():
+    """Example test for ``length_batch_input_outputs``."""
+    sequences = ["ASDF", "GHJKL", "PILKN"]
+    seqs_batches, seq_lens = length_batch_input_outputs(sequences)
+    assert len(seqs_batches) == len(set([len(x) for x in sequences]))
+    assert len(seq_lens) == len(set([len(x) for x in sequences]))
+
+
+def test_evotuning_pairs():
+    """Unit test for evotuning_pairs function."""
+    sequence = "ACGHJKL"
+    x, y = evotuning_pairs(sequence)
+    assert x.shape == (len(sequence) + 1, 10)  # embeddings ("x") are width 10
+    assert y.shape == (len(sequence) + 1, 25)  # output is one of 25 chars
